Added Specular Lighting

+++ Light.cpp
@@ -19,8 +19,8 @@ GLfloat xDirection, GLfloat yDirection, GLfloat zDirection, GLfloat dIntensity)
     diffuseIntensity = dIntensity;
 }
 
-void Light::useLight(GLfloat ambientIntensityLocation, GLfloat ambientColourLocation,
-    GLfloat diffuseIntensityLocation, GLfloat directionLocation)
+void Light::useLight(GLuint ambientIntensityLocation, GLuint ambientColourLocation,
+    GLuint diffuseIntensityLocation, GLuint directionLocation)
 {
     glUniform3f(ambientColourLocation, colour.x, colour.y, colour.z);
     glUniform1f(ambientIntensityLocation, ambientIntensity);
